# src/Prdb.Viewer.Host/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "prdb-viewer",
    "description": "The versioned API used by the prdb-viewer browser application.",
    "version": "v1"
  },
  "paths": {
...
    "/api/access/bootstrap": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BootstrapRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BootstrapResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/sign-in": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignInRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/registration-requests": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegistrationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegistrationRequestResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/access/recover": {
      "post": {
        "tags": [
          "Access"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecoverRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoverResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/accounts/{accountId}/approve": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{accountId}/disable": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountActionResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/admin/accounts/{accountId}/recovery-code": {
      "post": {
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "name": "accountId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RecoveryCodeResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/personal/playback-attempts/{playbackAttemptId}/end": {
      "post": {
        "tags": [
          "Personal State"
        ],
        "parameters": [
          {
            "name": "playbackAttemptId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EndPlaybackAttemptResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AccountActionResponse": {
        "required": [
          "verdict"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/AccountActionVerdict"
          }
        }
      },
      "AccountActionVerdict": {
        "enum": [
          "Completed",
          "NotFound",
          "InvalidState",
          "LastAdministrator"
        ]
      },
      "AccountAuthority": {
        "enum": [
          "User",
          "Administrator"
        ]
      },
...
      "BootstrapClaimVerdict": {
        "enum": [
          "Created",
          "InvalidInput",
          "InvalidAuthorization",
          "AlreadyClaimed"
        ]
      },
      "BootstrapRequest": {
        "required": [
          "authorization",
          "username",
          "password",
          "email"
        ],
        "type": "object",
        "properties": {
          "authorization": {
            "type": [
              "null",
              "string"
            ]
          },
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "BootstrapResponse": {
        "required": [
          "verdict",
          "account"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/BootstrapClaimVerdict"
          },
          "account": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SignedInAccountResponse"
              }
            ]
          }
        }
      },
...
      "EndPlaybackAttemptResponse": {
        "required": [
          "ended"
        ],
        "type": "object",
        "properties": {
          "ended": {
            "type": "boolean"
          }
        }
      },
...
      "RecoverRequest": {
        "required": [
          "username",
          "recoveryCode",
          "newPassword"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "recoveryCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "newPassword": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RecoverResponse": {
        "required": [
          "verdict"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/RecoveryVerdict"
          }
        }
      },
      "RecoveryCodeResponse": {
        "required": [
          "verdict",
          "recoveryCode",
          "expiresAt"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/AccountActionVerdict"
          },
          "recoveryCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "expiresAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          }
        }
      },
      "RecoveryVerdict": {
        "enum": [
          "PasswordReplaced",
          "InvalidInput",
          "InvalidCode"
        ]
      },
      "RegistrationRequest": {
        "required": [
          "username",
          "password",
          "email"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "RegistrationRequestResponse": {
        "required": [
          "verdict"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/RegistrationRequestVerdict"
          }
        }
      },
      "RegistrationRequestVerdict": {
        "enum": [
          "Submitted",
          "InvalidInput",
          "InstallationUnclaimed"
        ]
      },
...
      "SignedInAccountResponse": {
        "required": [
          "id",
          "username",
          "email",
          "authority",
          "csrfToken"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "authority": {
            "$ref": "#/components/schemas/AccountAuthority"
          },
          "csrfToken": {
            "type": "string"
          }
        }
      },
      "SignInRequest": {
        "required": [
          "username",
          "password"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": [
              "null",
              "string"
            ]
          },
          "password": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "SignInResponse": {
        "required": [
          "verdict",
          "account"
        ],
        "type": "object",
        "properties": {
          "verdict": {
            "$ref": "#/components/schemas/SignInVerdict"
          },
          "account": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/SignedInAccountResponse"
              }
            ]
          }
        }
      },
      "SignInVerdict": {
        "enum": [
          "SignedIn",
          "InvalidCredentials",
          "ApprovalPending",
          "Disabled"
        ]
      },
...
    }
  },
...
}